update workflow diagram and refactor components strcut
