+++ src/components/common/drawers/index.ts
@@ -1,0 +1,3 @@
+import InfoDrawer from './InfoDrawers';
+
+export { InfoDrawer };
