+++ src/components/common/spinner/index.tsx
@@ -1,0 +1,4 @@
+import Spinner from './Spinner';
+import SpinnerWithText from './SpinnerWithText';
+
+export { Spinner, SpinnerWithText };
