+++ src/components/diagrams/sections/cttsoDiagram.tsx
@@ -70,12 +70,61 @@ export const nodes = [
       ],
       detail: {
         status: 'SUCCEEDED',
-        description: 'this is a description',
+        description: 'Bcl convert succeded',
       },
     },
     draggable: true,
     selectable: true,
     position: { x: 600, y: 100 },
+  },
+  {
+    id: '4',
+    type: 'workflow',
+    data: {
+      type: 'TSO_CTDNA_TUMOR_ONLY',
+      title: 'TSO_CTDNA_TUMOR_ONLY',
+      handlers: [
+        {
+          type: 'target',
+          position: Position.Left,
+          id: 'TSO_CTDNA_TUMOR_ONLY-handle-0',
+        },
+        {
+          type: 'source',
+          position: Position.Right,
+          id: 'TSO_CTDNA_TUMOR_ONLY-handle-1',
+        },
+      ],
+      detail: {
+        status: 'SUCCEEDED',
+        description: 'TSO_CTDNA_TUMOR_ONLY succeded',
+      },
+    },
+    draggable: true,
+    selectable: true,
+    position: { x: 950, y: 100 },
+  },
+  {
+    id: '5',
+    type: 'workflow',
+    data: {
+      type: 'UMCCRISE',
+      title: 'UMCCRISE',
+      handlers: [
+        {
+          type: 'target',
+          position: Position.Left,
+          id: 'UMCCRISE-handle-0',
+        },
+      ],
+      detail: {
+        status: 'SUCCEEDED',
+        description: 'UMCCRISE succeded',
+      },
+    },
+    draggable: true,
+    selectable: true,
+    position: { x: 1300, y: 100 },
   },
 ] satisfies Node[];
 
@@ -96,6 +145,26 @@ export const edges = [
     type: 'smoothstep',
     sourceHandle: 'BSSH-handle-0',
     targetHandle: 'BCL_CONVERT-handle-0',
+    animated: true,
+    markerEnd: 'arrow',
+  },
+  {
+    id: 'e3-4',
+    source: '3',
+    target: '4',
+    type: 'smoothstep',
+    sourceHandle: 'BCL_CONVERT-handle-1',
+    targetHandle: 'TSO_CTDNA_TUMOR_ONLY-handle-0',
+    animated: true,
+    markerEnd: 'arrow',
+  },
+  {
+    id: 'e4-5',
+    source: '4',
+    target: '5',
+    type: 'smoothstep',
+    sourceHandle: 'TSO_CTDNA_TUMOR_ONLY-handle-1',
+    targetHandle: 'UMCCRISE-handle-0',
     animated: true,
     markerEnd: 'arrow',
   },
