+++ src/components/stats/index.tsx
@@ -1,0 +1,3 @@
+import StatsCards from './StatsCards';
+
+export { StatsCards };
