+++ src/components/tables/index.tsx
@@ -1,5 +1,4 @@
 import Table from './Table';
 import GroupedTable from './GroupedRowTable';
 
-export default Table;
-export { GroupedTable };
+export { Table, GroupedTable };
