+++ src/pages/Metadata/components/SubjectTable.tsx
@@ -1,6 +1,6 @@
 import { useMetadataFullSubjectModel } from '@/api/metadata';
 import { components } from '@/api/types/metadata';
-import Table from '@/components/tables';
+import { Table } from '@/components/tables';
 import { Column } from '@/components/tables/Table';
 import { classNames } from '@/utils/utils';
 import { MagnifyingGlassIcon } from '@heroicons/react/20/solid';
