+++ src/pages/Metadata/index.tsx
@@ -1,8 +1,8 @@
-import Card from '@/components/layouts/cards';
+import { Card } from '@/components/layouts/cards';
 import { Suspense } from 'react';
 import { SubjectTable } from './components/SubjectTable';
-import DetailedErrorBoundary from '@/components/common/error';
-import SpinnerWithText from '@/components/common/spinnerWithText';
+import { DetailedErrorBoundary } from '@/components/common/error';
+import { SpinnerWithText } from '@/components/common/spinner';
 
 export default function MetadataPage() {
   return (
