+++ src/stories/Card.stories.tsx
@@ -1,4 +1,4 @@
-import Card from '@/components/layouts/cards';
+import { Card } from '@/components/layouts/cards';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta: Meta = {
